+++ test/unit/lib/convert-symbols-to-slack-emoji.js
@@ -7,6 +7,10 @@ describe('convertSymbolsToSlackEmoji', function () {
     expect(convertSymbolsToSlackEmoji('Foo {t} Bar')).to.equal('Foo :mana-t: Bar')
   })
 
+  it('replaces hybrid mana symbols with mana named emojis', function () {
+    expect(convertSymbolsToSlackEmoji('Foo {u/b} Bar')).to.equal('Foo :mana-ub: Bar')
+  })
+
   it('leaves text without symbols unchanged', function () {
     expect(convertSymbolsToSlackEmoji('Foo Bar')).to.equal('Foo Bar')
   })