+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 
 * Support strings converting mana symbols ({m}) to Slack emoji syntax (:mana-m:)
+* Support strings converting mana symbols ({m}) to Discord emoji syntax (:manam:)
 
 _Breaking Changes_
 * Module is now instantiated as an instance